+++ plugins/frontend/backstage-plugin-datadog/src/alpha/index.ts
@@ -1,0 +1,3 @@
+export * from './apis';
+export * from './entityCards';
+export * from './entityContent';