+++ plugins/backend/rag-ai-backend-retrieval-augmenter/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @roadiehq/rag-ai-backend-retrieval-augmenter
 
+## 0.3.6
+
+### Patch Changes
+
+- f040722: Fixed missing metadata from TechDocs embeddings
+
+## 0.3.5
+
+### Patch Changes
+
+- c0ae442: Add `title` and `location` to metadata of TechDocs embeddings
+
+## 0.3.4
+
+### Patch Changes
+
+- f5bab23: Upgraded the langchain dependencies of the rag-ai plugins
+- Updated dependencies [f5bab23]
+  - @roadiehq/rag-ai-node@0.1.4
+
 ## 0.3.3
 
 ### Patch Changes