+++ plugins/frontend/backstage-plugin-prometheus/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @roadiehq/backstage-plugin-prometheus
 
+## 2.11.0
+
+### Minor Changes
+
+- f4a5909: Allow the title to be set on the prometheus entity card.
+
+## 2.10.0
+
+### Minor Changes
+
+- 6546676: Allow setting the query on the card.
+
 ## 2.9.0
 
 ### Minor Changes